Use the gem root to load provided default files
This has only been working incidentally so far, because all of my tests
had been run within the context of the working tree on my filesystem.
When I tried initializing a new weblog it became clear to me I needed to
reference the installed gem root directory for files provided by the
gem.