Make `SimpleHTMLFromERB` support actual ERB files
Before this commit this builder didn't quite work. Now we can build HTML
from ERB files and even provide the ERB files with information about the
current Lifer project. See the docs in the commit for more info.

Biggest change here is that we no longer use `<%= yield %>` and `<%=
content %>` is a special variable for layout files from now on.