Fix infinite loops when requiring Ruby files
Lifer provides a feature to load user-provided Ruby files. One of the
nicest things about this feature is to add Lifer selections and custom
Lifer settings, i.e.:

    Lifer.register_setting :my_beautiful_yaml_key

Unfortunately, the above example was resulting in an infinite loop. This
is because there's an implicit reference to the sole instance of
`Lifer::Brain` before the `@@brain` is memoized. This commit works
around the issue by simply loading user-provided files later, after
we're 100% sure that the brain has been initialized and memoized.

Co-authored-by: Chris Todorov <chris@super.gd>