Let layouts specify a parent layout
This gives us more flexibility when dealing with functional HTML layouts
vs. content layouts. To better explain:

  -> base-layout.html.liquid
  ---> layout-for-blog-posts.html.liquid
  -----> entry.html.md

Now, `layout-for-blog-posts.html.liquid can inherit the `<head>` and
`<body>` contents from the base layout. It can be focused on the content
 layout rather than the nitty gritty.

This is sort of like Jekyll's "layouts can have YAML frontmatter" but
much smaller in scope. Because of how Liquid works, though, the `{%
layout %}` block tag is a bit of a hack. I wanted it to be like a render
block, but that's not really a thing I can provide as far as I can tell.