Handle even more prebuilder errors
While manually testing, I saw an issue where a command errored out, but
Lifer kept running as normal. To resolve, I we had to manually check
`stderr` for error messaging and raise on that.

This might not be the perfect solution, but I think it's better than it
was before.