Improve test run speeds
We can use `parallel_tests` to increase the speed of test runs.

There is one tests that takes 3 seconds because it has to wait for Puma
to (reliably) start. We can skip this locally for now and only run it on
CI. But maybe in the future we should just delete that test?