Improve HTML entry "titles"
HTML entries currently have no way of providing metadata about
themselves, so for now we extract the title from the filename. However,
"index" is not a descriptive title, so if the file's permalink indicates
that it's an index file, we should go up a level and take the previous
"non-index" basename.