Ensure collection entries are output in an order
Before this, there was no telling what order a collection's entries
collection would be output in. This doesn't make sense in a lot of
scenarios--like for blogs. To resolve this, I decided we should always
output entries in an order.

Unfortunately, this made me realize that HTML entries have no date
metadata attached to them. Making Lifer builds output errors when
building HTML entries in "the right order". This is especially a problem
because Markdown and HTML entries can co-exist in a single collection.

The solution for now is to give HTML entries a default date. And that
default date is much older than anything else that could be in the
collection. But in the long term, I think we need a way to let HTML
documents provide metadata about themselves. I wonder if there's a nice
way to do this without putting weird garbage or frontmatter in them.