Introduce new method for inline integration tests
After some use, the original `Files` helper methods are getting a bit
convoluted. One other problem with them is that the depend on the
committed test Lifer project, which keeps getting bigger and more
convoluted and harder to maintain over time.

The benefit of this is that you can create a realistic-looking Lifer
project with a single instance of a class, and the Lifer project is
just a temporary directory.