Support Liquid environments (`liquid-5.6.0`)
Liquid has introduced the concept of environments. When parsing
documents, you now pass your parser a local, custom environment instead
of using a global, custom environment. This is good. I needed to make
changes to support this, though, because the changes were breaking.

This commit collects the required changes. You can see that this means
registering filters, tags, and a file system *within* the built
environment.