Provide `{% layout %}` all environment context
If a Liquid template using the `layout` tag also happened to contain a
`render` tag, i.e.

    {% layout "my_layout" %}

    <article>{{ content }}</article>

    {% render "footer" %}

The rendered partial would be missing a reference to the Liquid
environment and other document context accessible to all other templates
and entries being rendered. Because this was a bit difficult to track
down, I ensured existing tests cover this case.