Improve validity of RSS output
Obviously Lifer can only do so much to make RSS feeds valid. Some
invalidities could be due to the content of an RSS feed item, which is
out of our hands. However, in testing some simple output, it became
clear to me that we weren't doing enough to ensure there's always valid
`<managingEditor>` data.

This commit ensures that more RSS feeds will be valid out-of-the-box. :)