Allow non-ASCII characters in entry summaries
This was a bug. When I originally implemented summaries as first
paragraphs, I didn't realize that using the Kramdown representation of
the document would result in special characters being transformed badly:

    In the ldquorealmrdquo where dreams dance

This change ensures that we don't end up with trash like that, or other
HTML trash.

The summary is meant to be ideal for things like <meta> description
tags, which should only contain plain text.